feat: removed constructor restrictions for Unit and Dimension
Originally, the constructors for Unit and Dimension were private. The
derived classes BaseUnit and BaseDimension provided constructors. The
idea was to only allow the user to build units from base units.

However, this ended up being an issue for designing the UnitRegistry,
because we needed to be able to default construct a unit for the parser.

So, this commit moves the constructors into public.