docs: added writeup about unit conversions for offset units
I thought we had a bug related to conversions of offset units, but it
turned out it was a misunderstanding about how the unit should be
defined.

When converting a kelvin temperature to rankine, we use

R = 9 K / 5

but really this is a conversion for the numerical value of the
temperature quantity. We should write it as

T_R = 9 T_K / 5

The correct conversion between *units* is

900 R = 500 K

or

R = 5 K /9